Add support and tests for spatial rule 21655. This necessitated changing the SampledField implementation extensively, as that stored everything as ints, instead of doubles. Nothing in the spec seemed to indicate that the sampledField could only have ints, so I had to change it to store doubles everywhere instead. At that point, I could check for floats.

+++ src/sbml/packages/spatial/extension/test/TestReadSpatialExtension.cpp
@@ -26,7 +26,7 @@ LIBSBML_CPP_NAMESPACE_USE
   /** @endcond doxygenIgnored */
 
 
-std::string dataToString(int* field, int numSamples1, int length)
+std::string dataToString(double* field, int numSamples1, int length)
 {
   stringstream builder;
   for (int i = 0; i < length; ++i)
@@ -655,13 +655,13 @@ END_TEST
   
   // test new API 
   int length1 = field->getUncompressedLength();
-  int* array1 = new int[length1]; 
+  double* array1 = new double[length1]; 
   fail_unless(length1 == 3591);
   field->getUncompressed(array1);
   string test1 = dataToString(array1, field->getNumSamples1(), length1);
   fail_unless(test1 == expected);
 
-  int* result; int resultLength;
+  double* result; size_t resultLength;
   field->getUncompressedData(result, resultLength);
 
   fail_unless(resultLength == length1);
@@ -674,7 +674,7 @@ END_TEST
   // test new API
   int uncompressed = field->getUncompressedLength();
   fail_unless(resultLength == uncompressed);
-  int* more = new int[uncompressed]; 
+  double* more = new double[uncompressed]; 
   field->getUncompressed(more);
   resultString = dataToString(more, field->getNumSamples1(), resultLength);
   fail_unless(resultString == expected);
@@ -773,13 +773,13 @@ END_TEST
   
   // test new API 
   int length1 = field->getUncompressedLength();
-  int* array1 = new int[length1]; 
+  double* array1 = new double[length1];
   fail_unless(length1 == 3591);
   field->getUncompressed(array1);
   string test1 = dataToString(array1, field->getNumSamples1(), length1);
   fail_unless(test1 == expected);
 
-  int* result; int resultLength;
+  double* result; size_t resultLength;
   field->getUncompressedData(result, resultLength);
 
   fail_unless(resultLength == length1);
@@ -792,7 +792,7 @@ END_TEST
   // test new API
   int uncompressed = field->getUncompressedLength();
   fail_unless(resultLength == uncompressed);
-  int* more = new int[uncompressed]; 
+  double* more = new double[uncompressed]; 
   field->getUncompressed(more);
   resultString = dataToString(more, field->getNumSamples1(), resultLength);
   fail_unless(resultString == expected);
